MAG-403: ApplePay - Add config for PDP
