+++ Block/ApplePay/AbstractButton.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace Payplug\Payments\Block\ApplePay;
 
 use Magento\Bundle\Model\Product\Type as BundleProductType;
-use Magento\Catalog\Api\Data\ProductInterface;
 use Magento\Checkout\Model\Session as CheckoutSession;
 use Magento\Framework\Exception\LocalizedException;
 use Magento\Framework\Exception\NoSuchEntityException;
@@ -18,31 +17,19 @@ use Payplug\Payments\Gateway\Config\ApplePay as ApplePayMethodCode;
 use Payplug\Payments\Helper\ApplePay;
 use Payplug\Payments\Model\Payment\ApplePay\ConfigProvider;
 
-class Button extends Template
+abstract class AbstractButton extends Template
 {
     public function __construct(
-        private readonly ApplePay $applePayHelper,
         private readonly ConfigProvider $configProvider,
         private readonly Json $jsonSerializer,
         private readonly StoreManagerInterface $storeManager,
         private readonly CheckoutSession $checkoutSession,
-        private readonly Registry $registry,
+        protected readonly ApplePay $applePayHelper,
+        protected readonly Registry $registry,
         Context $context,
         array $data = []
     ) {
         parent::__construct($context, $data);
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function _toHtml(): string
-    {
-        if ($this->applePayHelper->canDisplayApplePay() && $this->applePayHelper->canDisplayApplePayOnCart()) {
-            return parent::_toHtml();
-        }
-
-        return '';
     }
 
     public function getConfig(): array
@@ -74,11 +61,6 @@ class Button extends Template
         }
 
         return false;
-    }
-
-    public function getProduct(): ?ProductInterface
-    {
-        return $this->registry->registry('product');
     }
 
     private function getCurrencyCode(): string
