+++ view/frontend/layout/checkout_cart_index.xml
@@ -12,7 +12,7 @@
         </referenceContainer>
 
         <referenceContainer name="checkout.cart.methods">
-            <block class="Payplug\Payments\Block\ApplePay\Button" name="checkout.cart.methods.button.apple.pay" template="Payplug_Payments::cart/button-apple-pay.phtml" after="-" />
+            <block class="Payplug\Payments\Block\ApplePay\CartButton" name="checkout.cart.methods.button.apple.pay" template="Payplug_Payments::cart/button-apple-pay.phtml" after="-" />
         </referenceContainer>
     </body>
 </page>
