+++ view/frontend/templates/product/button-apple-pay.phtml
@@ -3,11 +3,11 @@
 declare(strict_types=1);
 
 use Magento\Framework\Escaper;
-use Payplug\Payments\Block\ApplePay\Button;
+use Payplug\Payments\Block\ApplePay\ProductButton;
 
 /**
  * @var Escaper $escaper
- * @var Button $block
+ * @var ProductButton $block
  */
 ?>
